Fix/26x (#82)
* added DTI big endian message sending & changed dim to old pinout temporarily

* Auto-generate symbol files [skip ci]

* tssi testing

* added front-rear bias, changed torque-current mapping

* Auto-generate symbol files [skip ci]

* adding sensoric vel

* thermistor transfer functions and cell balancing hysteresis

* fix transfer function a little

* disable cell balance, added back caninit for lvb, changed hv board num to 17

* fixed motors code - not send in hv en, temp dssi code for testing

* Auto-generate symbol files [skip ci]

* added back hvc errors, removed dssi

* fast/slow mode working, used current control bc of encoder issue, added back hvc errors on vsm and fixed new pins

* displaying sensoric vel on dim screen

* Auto-generate symbol files [skip ci]

* added sensoric vel on screen, need to change config can ids

* Auto-generate symbol files [skip ci]

* updated swangle adc transfer function

* debugging dcm errors

* calculating temps

* cell temps working

* Auto-generate symbol files [skip ci]

* got rid of bms

* testing ebs, transfer function for ebs rpressure

* Auto-generate symbol files [skip ci]

* steering motor code on updated dcm

* Auto-generate symbol files [skip ci]

* Corrected inspection mission code

* Auto-generate symbol files [skip ci]

* added back dssi code

* setting power limit w bias in fast/test gear

* 35/65 torque bias

* Not building on my end

* Auto-generate symbol files [skip ci]

* fixed can driver issues

* Auto-generate symbol files [skip ci]

* added sensoric dist & vel to dim screen

* Auto-generate symbol files [skip ci]

* fixing build errors

* Auto-generate symbol files [skip ci]

* dim gpio pins corrected

* added back cell temps

* include for fabs

* remapped tpos

* fixing inspection code & swangle transfer function

* fixed swangle transfer function

* inspection mission working

* inspection mission working

* Auto-generate symbol files [skip ci]

---------

+++ DIM/Inc/state.h
@@ -62,7 +62,7 @@ cmr_canDVMode_t stateGetDVReq(void);
 
 uint32_t get_test_message_id();
 
-float getSpeedKmh(void);
+uint8_t getSpeedKmh(void);
 
 float getOdometer(void);
